🗺️ Atlas: Extract Native Module
